Change single-page to multi-page - includes minor fix to Dashboard and naming
