+++ client/package.json
@@ -16,7 +16,6 @@
     "luxon": "^1.22.2",
     "vue": "^2.6.11",
     "vue-chartjs": "^3.5.0",
-    "vue-router": "^3.1.5",
     "vuetify": "^2.2.11"
   },
   "devDependencies": {
