+++ client/src/pages/dashboard/Dashboard.vue
@@ -1,60 +1,61 @@
 <template>
-  <v-row>
-    <v-col xs="12">
-      <v-tabs
-        v-model="activeZoneGroup"
-        grow
-        show-arrows>
-        <v-tabs-slider />
+  <v-app>
+    <v-content>
+      <v-container fill-height fluid>
+        <v-tabs
+          v-model="activeZoneGroup"
+          grow
+          show-arrows>
+          <v-tabs-slider />
 
-        <v-tab
-          v-for="group in zoneGroups"
-          :key="group.id">
-          {{ group.name }}
-        </v-tab>
-        <v-tabs-items v-model="activeZoneGroup">
-          <v-tab-item v-for="group in zoneGroups" :key="group.id">
-            <v-row class="mb-6">
-              <v-col class="text-center">
-                <img :src="group.layout" v-if="group.layout !== null" />
-                <img src="https://via.placeholder.com/728x90?Text=placeholder" v-else/>
-              </v-col>
-            </v-row>
-            <v-row>
-              <v-col
-                v-for="zone in group.zones"
-                :key="zone.id"
-                :lg="4"
-                :md="6"
-                :sm="12">
-                <ZoneButton
-                  @click="openZoneDialog(zone)"
-                  :name="zone.name"
-                  :people-count="zone.persons_count"
-                  :alert-unauthorized="zone.alertUnauthorized"
-                  :alert-unknown-person="zone.alertUnknownPerson"
-                  :alert-overstay="zone.alertOverstay"/>
-              </v-col>
-            </v-row>
-          </v-tab-item>
-        </v-tabs-items>
-      </v-tabs>
-    </v-col>
-    <ZoneDialog
-      v-model="dialog"
-      :zone="activeZone"
-      @people-updated="handlePeopleUpdated"
-      @alerts-updated="handleAlertsUpdated"
-      />
-  </v-row>
+          <v-tab
+            v-for="group in zoneGroups"
+            :key="group.id">
+            {{ group.name }}
+          </v-tab>
+          <v-tabs-items v-model="activeZoneGroup">
+            <v-tab-item v-for="group in zoneGroups" :key="group.id">
+              <v-row class="mb-6">
+                <v-col class="text-center">
+                  <v-img :src="group.layout" v-if="group.layout !== null" contain/>
+                  <img src="https://via.placeholder.com/728x90?Text=placeholder" v-else/>
+                </v-col>
+              </v-row>
+              <v-row>
+                <v-col
+                  v-for="zone in group.zones"
+                  :key="zone.id"
+                  :lg="4"
+                  :md="6"
+                  :sm="12">
+                  <ZoneButton
+                    @click="openZoneDialog(zone)"
+                    :name="zone.name"
+                    :people-count="zone.persons_count"
+                    :alert-unauthorized="zone.alertUnauthorized"
+                    :alert-unknown-person="zone.alertUnknownPerson"
+                    :alert-overstay="zone.alertOverstay"/>
+                </v-col>
+              </v-row>
+            </v-tab-item>
+          </v-tabs-items>
+        </v-tabs>
+        <ZoneDialog
+          v-model="dialog"
+          :zone="activeZone"
+          @people-updated="handlePeopleUpdated"
+          @alerts-updated="handleAlertsUpdated"
+          />
+      </v-container>
+    </v-content>
+  </v-app>
 </template>
 
 <script>
-// @ is an alias to /src
 import _ from 'lodash'
 import api, { ALERT_TYPES } from '@/api'
-import ZoneButton from '@/components/ZoneButton'
-import ZoneDialog from '@/components/ZoneDialog'
+import ZoneButton from './components/ZoneButton'
+import ZoneDialog from './components/ZoneDialog'
 
 export default {
   name: 'Dashboard',
