+++ client/src/pages/dashboard/components/ZoneButton.vue
