+++ client/src/pages/dashboard/components/ZonePeopleView.vue
@@ -9,11 +9,18 @@
         <v-btn color="primary" class="mb-2" @click="refreshData" :disabled="loading">Refresh</v-btn>
       </v-toolbar>
     </template>
+    <template v-slot:item.name="{ item }">
+      <span v-if="item.name">{{ item.name }}</span>
+      <span v-else>UNKNOWN</span>
+    </template>
     <template v-slot:item.from="{ item }">
       <span>{{ (new Date(item.from)).toLocaleString('en-US', dateTimeFormatOptions) }}</span>
     </template>
-    <template v-slot:item.picture="{ item }">
-      <img :src="item.picture"/>
+    <template v-slot:item.avatar="{ item }">
+      <v-avatar size="48">
+        <v-img v-if="item.avatar !== null" :src="item.avatar"/>
+        <v-img v-else src="/avatar-placeholder.png" />
+      </v-avatar>
     </template>
     <template v-slot:item.remarks="{ item }">
       <span v-if="item.alerts.length === 0">-</span>
@@ -36,7 +43,7 @@ import { DATE_TIME_FORMAT } from '@/common/date-time'
 
 const headers = [
   { text: 'ID', value: 'id' },
-  { text: 'Foto', value: 'picture', sortable: false },
+  { text: 'Foto', value: 'avatar', sortable: false },
   { text: 'Nama', value: 'name' },
   { text: 'Waktu Masuk', value: 'from' },
   { text: 'Keterangan', value: 'remarks' },
