+++ client/src/pages/zone-people-hourly-count/components/HourlyPeopleChart.vue
