+++ client/src/pages/zone-people-hourly-count/index.js
@@ -1,0 +1,10 @@
+import Vue from 'vue'
+import ZonePeopleHourlyCount from './ZonePeopleHourlyCount.vue'
+import vuetify from '@/plugins/vuetify'
+
+Vue.config.productionTip = false
+
+new Vue({
+  vuetify,
+  render: h => h(ZonePeopleHourlyCount)
+}).$mount('#app')
