+++ client/src/pages/zone-settings/index.js
@@ -1,0 +1,10 @@
+import Vue from 'vue'
+import ZoneSettings from './ZoneSettings.vue'
+import vuetify from '@/plugins/vuetify'
+
+Vue.config.productionTip = false
+
+new Vue({
+  vuetify,
+  render: h => h(ZoneSettings)
+}).$mount('#app')
