--- client/src/router/index.js
@@ -1,48 +1,0 @@
-import Vue from 'vue'
-import VueRouter from 'vue-router'
-
-import Home from '../views/Home.vue'
-import Dashboard from '../views/Dashboard.vue'
-import ZoneDailyReport from '../views/ZoneDailyReport'
-import ZoneDailyPeopleChart from '../views/ZoneDailyPeopleChart'
-
-Vue.use(VueRouter)
-
-const routes = [
-  {
-    path: '/',
-    name: 'Home',
-    component: Home
-  },
-  {
-    path: '/dashboard',
-    name: 'Dashboard',
-    component: Dashboard
-  },
-  {
-    path: '/daily-report',
-    name: 'ZoneDailyReport',
-    component: ZoneDailyReport
-  },
-  {
-    path: '/daily-people-chart',
-    name: 'ZoneDailyReport',
-    component: ZoneDailyPeopleChart
-  },
-  {
-    path: '/about',
-    name: 'About',
-    // route level code-splitting
-    // this generates a separate chunk (about.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "about" */ '../views/About.vue')
-  }
-]
-
-const router = new VueRouter({
-  mode: 'history',
-  base: process.env.BASE_URL,
-  routes
-})
-
-export default router
